Correct GEM_PATH and GEM_HOME setup
GEM_PATH and GEM_HOME were being joined through a simple concatenation. If root does not end in a / this fails (which is the case when using iron_worker run). Using File.join ensures the / is added if needed.

Also any system gems were not being included if ENV['GEM_PATH'] was empty. Some environments will always have GEM_PATH set to the proper paths. But often it is empty an we are just relying on the RubyGems defaults (i.e. ENV['GEM_PATH'] is empty but Gem.path is not). Updated the code to use Gem.path instead of ENV['GEM_PATH'] so system gems are used as a fallback (useful if running locally but doing a remote build and therefore there is not a __gems__ directory or if doing a local build but some gems have native extensions and are skipped).